make error refs consistent between + and the other binary operators
Summary:
all of the binary operators except for + use `Primitive.at loc` (e.g. `NumT.at loc`) as the result type of the expression. `at loc` sets the reason's `annot_loc` to `loc`, and then the annot loc is used as the reference location when that type errors.

So if you do `let x = 1 * 2; x.length;`, it says that `length` can't be found on `Number`, where `Number` is a ref pointing at `1 * 2`.

But currently if you replace `*` with `+`, `Number` is a ref pointing at the `declare class Number` in core.js. This happens because we do `NumT.why reason` instead of `NumT.at loc`, and the former doesn't set the annot loc.

Before and after on the new tests:
```
 --- arith.exp    2018-06-21 07:33:19.277833439 -0700
+++ arith.out    2018-06-21 07:33:20.663832942 -0700
@@ -344,13 +344,13 @@
 Cannot get `(x + y).length` because property `length` is missing in `Number` [1].

    Arith.js:93:5
-    93|     (x + y).length; // error
-            ^^^^^^^^^^^^^^
+   93|     (x + y).length; // error
+           ^^^^^^^^^^^^^^

 References:
-   <BUILTINS>/core.js:124:15
-   124| declare class Number {
-                      ^^^^^^ [1]
+   Arith.js:93:6
+   93|     (x + y).length; // error
+            ^^^^^ [1]

 Error ---------------------------------------------------------------------------------------------------- Arith.js:94:5
@@ -358,13 +358,13 @@
 Cannot get `(1 + 2).length` because property `length` is missing in `Number` [1].

    Arith.js:94:5
-    94|     (1 + 2).length; // error
-            ^^^^^^^^^^^^^^
+   94|     (1 + 2).length; // error
+           ^^^^^^^^^^^^^^

 References:
-   <BUILTINS>/core.js:124:15
-   124| declare class Number {
-                      ^^^^^^ [1]
+   Arith.js:94:6
+   94|     (1 + 2).length; // error
+            ^^^^^ [1]
```

Reviewed By: gabelevi

Differential Revision: D8567653

fbshipit-source-id: 9fa66fea753183a216cf752be8027578bc3367de