Typed AST: move error AST nodes to their own module
Summary: When generating the (not yet typed) typed AST sometimes an error prevents constraint generation from traversing the source AST further, which in turn prevents us from translating that portion of the source AST into a typed AST. In these situations, we return a bogus "error node". This diff puts all the error nodes in a centralized place so that it's easy to look up and modify their representations.

Reviewed By: panagosg7

Differential Revision: D8502529

fbshipit-source-id: 3379316dc10fea17fbbc83d50ff8f7f238830e4f