Combine DerivedInstance and DerivedStatics SuperT constraints
Summary:
Before this diff, we would emit two separate SuperT constraints when checking
classes: one for the instance type, one for statics.

This caused unusual behavior when the value in the extends position was not a
class -- we would error twice -- whch panagosg7 noticed in D7800039. This
eliminates the redundant error in that test.

This change also unblocks further simplification of class type checking.
Currently the SuperT constraints happen "on the side." I would like to change
this model to instead use SuperT to actually compute the class type. This would
involve adding a `tout` the SuperT, which would copy down properties and emit
an InstanceT. This is much simpler with a single SuperT constraint.

Reviewed By: panagosg7

Differential Revision: D7980558

fbshipit-source-id: 3010dea5d8174eafe420ba903b19f40864335cc8