Restructure class sig to defer super type construction
Summary:
Before this diff, we would construct types representing the super (and static
super) in `Class_sig.empty`. This diff keeps around the more structured
`Class_sig.super` data instead, only creating the types on demand.

This is a small step toward a larger goal. In follow-up diffs, I will change
the existing `SuperT` check constraint to actually construct the instance type.
It is useful to have the types in a structured form for this, so they can be
resolved in the correct order and with the appropriate semantics.

Reviewed By: panagosg7

Differential Revision: D8504933

fbshipit-source-id: b2f0686a784827c0e3f8b25f78cb411072f81870