Promote IdxWrapper to a def_t
Summary:
As an `internal_t`, `IdxWrapper` currently gets filtered out by `resolve_type`. Unfortunately this means that get-def, find-refs, and autocomplete return no results for property accesses on values of type `IdxWrapper`.

Because `IdxWrapper` is a type that can actually be assigned to an expression, it makes more sense as a `def_t` or maybe a `Type.t` than as an `internal_t`. I believe this is a more principled alternative to simply special-casing `IdxWrapper` in `Type.is_proper_def`, which is what `resolve_type` uses to filter out internal types.

Reviewed By: samwgoldman

Differential Revision: D8548165

fbshipit-source-id: 6c8096d072fc08c595c500d3b5c1f09b3aa1f3a9