Add some flavor to TypeT
Summary:
`TypeT` is used to define type aliases, but is also used for convenience in a few other
locations. E.g. when importing a class as a type:
```
import type { C } from 'M';
```
where `C` is a class in `M`.

Handling of `TypeT` may vary depending on each specific construction site. This diff adds
some extra information to the constructor, to be later used by the normalizer

Reviewed By: nmote

Differential Revision: D8245418

fbshipit-source-id: fd1a7e7042187deca361e3af286771533a356b3b