[Normalizer] Better reconstruction of type aliases (TypeT)
Summary:
[This commit is larger than usual, as it introduces a number of changes that lead to inconsistent
results if split up.]

One of the goals for the normalizer is to aggressively use nominal type information, whenever available,
instead of reconstructing large type structures. At the same time is needs to normalize the bodies of type
aliases (`TypeT`) through a one-off expansion of their type bodies.

The ground truth for type-alias names (from now on) will be `RTypeAlias`. This reason descriptor is
introduced in a single place in statement.ml. It is used to annotate the type alias itself (`TypeT`), as well
as the type body of the alias. So as a general rule of thumb (default behavior), whenever an `RTypeAlias`
is seen at the top-level of a type, then the respective type alias is return in its stead. Since pretty much
any type can appear under as the body of a type-alias, we trap cases like this before the actual type
pattern matching.

There are two cases where this process is circumvented:
1. The normalizer is configured to always expand aliases (through `opt_expand_type_aliases`).
2. We are resolving a "type-alias" type itself (ie. the input type is `TypeT`). In that case we resolve the
name of the alias with the above process and then compute the body of the type alias to get the following
structure:
```
type T    =    T_body
     ^         ^^^^^^
normal         one-off type alias
invocation     expansion
```
To enable this construction we introduce a field `under_type_alias` in `Env`, (defaulting to `None`), which
enables a *one-off* expansion of type-aliases. This means that it's only going to be set to `Some name`
for one-level of recursive calls, or until we hit a concrete constructor. In other words, it persists for constructors
like `OpenT` or `AnnotT` (plus some cases outlined in the caveats below).

The new `type_t` also uses the `kind` of the `TypeT`. E.g.:
* `TypeAliasKind`: Base case (with a small twist in comments)
* `ImportClassKind`: importing classes as types adds a `TypeT` wrapper around them. We still
want to reconstruct these types as classes instead of type-aliases. This handles this case.
* `ImportTypeofKind`, `ImportFunKind` use import-related reasons to extract the names of imported
types.

Reviewed By: samwgoldman

Differential Revision: D8246064

fbshipit-source-id: 1c6ddf4a30a8904c0911958131dbb9d5afec4755