Make Destructuring module return unit ASTs
Summary: The functions in the `typing/destructuring.ml` module are the only things that traverse patterns during constraint generation, so they will be the ones who generate pattern ASTs. This diff makes them return unit ASTs to get the main plumbing out of the way, and in the future these will be typed ASTs.

Reviewed By: panagosg7

Differential Revision: D8528786

fbshipit-source-id: d833bf6067a9ffc4287e8c83367574fc6b809514