Add tests for idx + unions
Summary:
The existing logic for this splits the union up, so every branch of the union
flows through the idx callback independently. This is safe, but creates a lot
of noise when there are errors.

In the added examples, we only need to emit a single error, but we end up
emitting three. Ideally, we would be able to say that the property look up
fails and propagate any. Doing so would require concretizing the union,
filtering nullish values, and packaging into a new union, being careful to
avoid non-termination due to recursion in the unions themselves.

Reviewed By: fishythefish

Differential Revision: D8618007

fbshipit-source-id: 9402d76996324dc51193ae1cbbef5aa7c781ac65