Rename Type.insttype fields_tmap/methods_tmap to own_props/proto_props
Summary:
This is just a rename. I have not changed any behaviors in this diff.

These names reflect the intended behavior here. It's useful to reflect the
intent in the names here, as I hope to solidify Flow's handling of property
layout in general.

That is, currently there is some confusion here. I will follow up with fixes,
which will make more sense with these new names.

Note that this naming is confusing for interfaces, which don't actually
distinguish between own and proto -- interfaces are flat. I plan to fix this by
separating interfaces from InstanceT in the future, so this temporary confusion
is OK.

Reviewed By: panagosg7

Differential Revision: D7856373

fbshipit-source-id: 347c35bf9f97484331534845a965ff412f322a84