[Normalizer] Only simplify type aliases of imported classes
Summary:
**Before:**
We were simplifying type aliases of all class types. For example,
```
// ex 1
class A {}
type T = typeof A;
```
the returned type for `T` would be simplified as:
```
Class(A, None)
```
The reason this shortcut was introduced in the first place was to simplify imports of classes:
```
// ex 2
import typeof { C } from 'm';
type _A = C;
```
Querying C in last line should still return the remote definition instead of the local one.

**Now:**
This changes the behavior for ex 1 to report `T` as
```
TypeAlias(T, Class(A, None))
```
The behavior of ex 2 remains the same, since there is not much point in reporting the import location
as provenance of `C`.

Reviewed By: samwgoldman

Differential Revision: D8664607

fbshipit-source-id: 18846a88dc6507716382db4fdb60331468a008f2