Global rename
Summary:
This is still very rough, as you can see from the numerous TODO items throughout. However, it does seem to work and it's a good starting point.

The overall goal is to produce a collection of (location, new_text) pairs which can be applied to the existing program text in order to correctly rename the symbol at the given location. This is the format for edits which Nuclide and the LSP both expect.

Right now, this can rename local variables and object properties, as demonstrated in the tests. Unfortunately it reprints the entire file, instead of giving a list of minimal edits. It is also not particularly safe, as no care is taken to avoid naming conflicts, nor do we make any effort to avoid renaming things which ought not to be renamed.

Here is the current design, at a high level:
* Run find references on the given location in order to get a list of locations which need to be changed, as well as information about what kind of symbol they are.
* For each file with references, get the AST for it and run an AST mapper which makes the correct substitutions based on the references in that file.
  * Note that due to complications like object shorthand and destructuring shorthand, we must do this as a pass over the AST rather than simply replacing the locations returned by find-refs with the new text.
  * Also due to object and destructuring shorthand, we need information about the nature of the reference to be available while we are walking the AST. See the implementation for details.
* Once we have the updated AST, we reprint it and return the start/end location along with the entire reprinted text.

Reviewed By: panagosg7

Differential Revision: D8595483

fbshipit-source-id: 5a2aa0dda0cae00eb2fb7016418ed8ba025c0ac6