Telemetry 1/8 - Typecheck produces summary of what it did; sends to monitor
Summary:
This diff is the first in a stack. I'll describe the architecture+motive for the stack.

In our large Hack project, we observed that 10% of requests from Nuclide are not being answered - roughly 3% because hh_server is stopped, 3% because it's busy initializing, 3% because it's in the middle of typechecking. It looks like about half our users suffer 15+ minutes of hh_server unavailability per day; some find it unavailable for 80%+ of their working day. I want to gather the same kind of end-to-end telemetry for Flow. One hypothesis is that Flow is worse than Hack because it has to deal with cycles and because it lacks saved-states. But we don't really know.

My aim is to get end-to-end telemetry about the user experience, conceived broadly enough that we can measure every possible source of failure.

WHAT METADATA SHOULD WE GATHER FOR EACH REQUEST?

Some requests from Nuclide will be rejected because the flow server is unavailable. Others will take too long to fulfill either because a typecheck is underway or just because there are too many others waiting to be dealt with first. Others might encounter internal failures to deliver results where they really should. And all the rest are success.

For the metadata to be actionable, we'd like to figure out, for failure cases, which of the above cases it was. I've set about logging the following metadata:

1. What was the request from Nuclide? (I want to record the original json request, rather than our parsed version of it, in case there are defects in the parser).
2. What was the state at the moment the request was received? Was the flow server even running? initializing? what was its state? what was the watcher state?
3. If the flow server did a typecheck before it could deal with this request, can we characterize that typecheck? How long did it take? How many dependent files? If cycles are indeed a major problem for flow, what was the biggest cycle encountered during this typecheck? (maybe in future we could also try to characterize all other requests ahead of this one in the queue).
4. How long was the request just waiting for the server to become ready to handle it? And once the server became ready, how long did the server take to handle it? How much client-side computation happened in lspCommand?
5. If there was an error, of course we want to know the error message and a stacktrace.

WHERE METADATA IS KNOWN?

The `flow lsp` process, in lspCommand.ml, is reliable and tenacious - we haven't observed any crashes, and Nuclide+VSCode ensure tenaciously that it's running. (if it does crash, then they'll automatically restart it, up to 5 times in 5 minutes). This makes it a good fall-back place to do logging.

Some information (the json of the request) is known in lspCommand.ml but not elsewhere.

Some information (the state of the server at the moment the request was received, and the time it was received) is known in lspCommand (if the server wasn't available) or in the monitor (if it was).

Some information (about typechecks that happened while the message was pending) are known in lspCommand and the monitor and the server.

Some information (about how the server performed in answering the request, and any errors that arose) is known in commandHandler.

So who should be responsible for gathering together all that metadata and logging it?

ARCHITECTURE

1. lspCommand gathers initial metadata about the request (when it arrived, what was status at the time it arrived)
2. The metadata is passed from lspCommand over to monitor+server, and thence back to lspCommand when commandHandler is finished. This is good because (A) we don't need to do any correlation, (B) should the server need to make multiple responses to a single request, it can decide which one of those responses is associated with it for logging.
3. While in commandHandler, metadata gathers any extra json_data, and also profiling information about how long it took on the server.
4. When back at lspCommand, metadata gathers information about any intervening commands that had finished (between the time this information arrived and when it was completed). Also the timing contribution on the client side.
5. lspCommand is responsible for logging all that metadata.

LOGGING FIELDS

0. **Event** - as always, this norm string is PERSISTENT_COMMAND_SUCCESS for either a successful response or an expected error response; or PERSISTENT_COMMAND_FAILURE for an unexpected error response (i.e. an uncaught exception, basically a coding bug). There are two new events: PERSISTENT_EXPECTED_ERROR and PERSISTENT_UNEXPECTED_ERROR for errors that arose spontaneously not associated with a request (e.g. connection termination).
1. **request** - this used to be a denorm string of the request, e.g. `didOpen path/file1.js`. I've changed it into a denorm JSON string, e.g. `{method:"didOpen", files: [...]}`. That'll let us better extract fields. For LSP requests, it will be the original JSON request.
2. **Timing** - as always, this is a json denorm of the abridged Profiling_js.finished structure that shows how the server fulfilled the request; `results.Profiling.wall.duration` shows how long the server took to execute the request (in seconds)
3. **Timing Client** - a new field, similar to Timing, but it's about time spent on the client. We hope this will be insignificant. If it's less than 5ms I don't bother logging it.
4. **Start Time, End Time, Time Taken Ms**. These fields used to be used for a few other events. Now I'm using them also for PERSISTENT_COMMAND_SUCCESS/FAILURE too. They will record from the time the request arrived at the monitor, until the time an answer was sent back to the client. (in the case of an lspCommand-only log, they show when it arrived at lspCommand until when it was responded to).
5. **data**. This used to be a norm json object storing request-specific stuff. Also for expected errors it stored `Error: <reason>` and for unexpected errors it stored `exn: <reason>`, and for both errors it stored `stack: <stack>`. I believe it should be *denorm*, so I changed it into that. And I stored some extra fields inside it (which will be absent if they're not available): `start_server_status`, `start_watcher_status`, `start_lsp_state`, `start_lsp_reason`, `delays`

----------

This diff is a small step in the journey. It makes it so that, when a typecheck completes, then the server gathers a short summary of what work was involved in the typecheck, and sends it on to the monitor.

Reviewed By: gabelevi

Differential Revision: D8250454

fbshipit-source-id: 4ba08d9eee4dd9bbf526588dd177b18c1b3ebe77