Telemetry 4b - retain metadata for each request
Summary:
This diff defines "metadata" - the extra information about each request which we'll want to log.

It modifies lspCommand so that metadata flows all the way through - initialized at the moment the request is received from the editor, send through to the monitor, and received back from the monitor.

It modifies commandHandler to write its part of the metadata.

The diff also refactors commandHandler to use Core_result rather than exceptions.

DESIGN NOTES:

I made "metadata" be a type with optional fields that gets updated along the way. I could instead have split it into three types "metadata_initially", "metadata_after_server", "metadata_after_back_at_lspCommand".

But that wouldn't have gained much. That's because the fields that server fills in are almost all optional (only exception is 'profiling'). In all, having three types for metadata seemed more confusing than one, without enough benefit.

I also made handle_persistent_unsafe return either an `IdeResponse (Ok|Error)`, or an `LspResponse (Ok|Error)`. That's because the meaning of "Ok" in the two worlds is different, and the meaning of "Error" in the two worlds is different...
* In the `IdeResponse` world, an Ok response means that handle_persistent_unsafe has already sent the message to the persistent client, and it's the job of handle_persistent to log it.
* In the `LspResponse` workd, an Ok response means that it's the job of handle_persistent to send the message to the persistent client, and it's the job of the client to log it.
* Similarly with Error.

They're so different that I thought they needed to be carefully separated.

Reviewed By: gabelevi

Differential Revision: D8473391

fbshipit-source-id: e4e4f3243c484a3e37508de0621659014fa04f3b