Telemetry 5b - lspCommand uses result not exceptions
Summary:
This is a small refactoring of lspCommand to use Core_result rather than exceptions. It makes some things easier.

Note: I didn't make the change completely. I still use `raise` in various places. That's because I just couldn't be bothered to explicitly gather callstacks in every single one of those places. It was easier and lazier to just use `raise` and to catch that and turn it into an Error. I'll defer the work of result-ifying it thoroughly to a future diff.

Reviewed By: gabelevi

Differential Revision: D8473395

fbshipit-source-id: c47630a4b32090df8a6968a83183777b1103cabf