Treat getters and setters as proto fields
Summary:
We're moving toward a more precise model of classes, where own and proto
properties are distinct. In reality, getters and setters live on the prototype,
but Flow treated them as "fields" instead.

The existing test for this had some assertions about shadowing. The new object
model solidifies what should happen in this instance, but we are only partially
moved to that model. I have just removed these assertions for now, but will add
comprehensive coverage as the new model falls into place.

Reviewed By: panagosg7

Differential Revision: D8667519

fbshipit-source-id: 3e42cc5a063008d3c588d127c773a92bbca12ea9