Make statement and toplevels return/throw unit ASTs
Summary:
Updates the `statement` and `toplevels` functions in `statement.ml` to produce unit ASTs. Sorry for the big diff.

Constraint generation of statements sometimes raises exceptions indicating abnormal control flow. Since an OCaml function can't simultaneously return and raise an exception, the abnormal control flow exceptions had to be updated to carry AST payloads. As the exception bubbles up through calls, the payload-carrying exception is caught, the payload AST is wrapped in to a bigger AST, and the exception is rethrown with this new payload.

If typed AST construction were to ever cause a meaningful performance hit, it would probably show up in this diff as this is the first time the entire program's AST is constructed without intermediately discarding AST nodes.

Things that still need to be done before moving onto adding types to the AST nodes:
 - currently DeclareFunction nodes without predicates are reconstructed as error nodes. Fixing this may require collecting some ASTs from the first pass through the statements, rather than exclusively from the second pass.
 - JSX expressions aren't translated

Reviewed By: panagosg7

Differential Revision: D8729790

fbshipit-source-id: 3ab7e351c0d806383610c45db33df10fec7a5c09