[Normalizer][type-tables] Rework tracking of tparams and reconstruction of BoundT
Summary:
**Motivation for this change**

Before, normalizing a type structure of the form:
```
ClassT (BoundT _)
```
would cause an exception in the normalizer and the position would be considered uncovered.

E.g.:
```
type C<T: R<any>> = Class<T>;
     ^
```

**How this diff addresses this**

1. The type tables now hold a richer `Type.typeparam` environment that includes information
about the upper bound of the `BoundT`.
2. This environment then also becomes part of the normalizer environment `env.tparams`.
3. When reconstructing `Ty.Class (Ty.Bound p)` we use the bound `b` for `p` (that can now
be recovered from `env.tparams`). Note here that this information was not available in
`Ty`types earlier.
4. We reconstruct `ClassT b`.

**Related changes that are included in this diff**

[I tried to isolate the following but some invariant was breaking causing some tests to fail.]

1. `Type.BoundT p` uses the information in `p` to reconstruct a `Ty.Bound _`. I believe this type
only appears annotating type parameter so there is no need to go through `lookup_tparam` in the
normalizer environment.
2. Because of the above change, tracking of type parameters has been simplified. In particular this
diff reverses some of the plumbing of D7401236. This introduces a lot of bulk to this change. The
meaningful changes are in ty_normalizer.ml.
3. Operations on type_tables do not go through `Env` anymore.

Reviewed By: samwgoldman

Differential Revision: D8597105

fbshipit-source-id: b24c780a6576c495aba0dbc96f1c2125109cd209