[Flow_js] Adds an exception for not_expect_bound
Summary:
Parameterizes the "BoundT not expected" assertion to not print a stacktrace in the log.
Motivation is that in the next diff we'll be using a callback in Flow_js from the normalizer
in which this exception is a valid result, so we would like to suppress the noise. The primary
case should not be affected.

This also exports `eval_type_destructor` to be used later by the normalizer.

Reviewed By: samwgoldman

Differential Revision: D8680939

fbshipit-source-id: 89a81d723716bacfadd6897501ba748609c83905