[Normalizer] Fix EvalT (type-destructor)
Summary:
There were two main reasons for this change:
1. @[5401580:samgoldman]'s suggestion in an earlier comment on this diff
2. I had been noticing several normalization failures (uncovered locs) to be due to `EvalT` handling.

The algorithm to reconstruct `EvalT (TypeDestructorT)` now is the following:
1. Lookup the `EvalT` in `Context.evaluated` in case this type has been evaluated before.
 If the search succeeds, then return the result. Otherwise go to 2.
2. Try to resolve the `EvalT` by issuing a call to `Flow_js.eval_destructor`. If this process terminates
without exceptions (it may cause new errors -- these will be ignored), we use the evaluated type
to resume normalization and return. Otherwise go to 3.
3. Some exception was thrown in step 2. This is (almost?) exclusively due to `EvalT`'s argument
containing "raw" `BoundT`s. In this case we cannot expect Flow to resolve this type, so we back out
and restore an *abstract* version of the `EvalT`, ie. one containing the destructor itself, for those
constructors that have an equivalent in the surface type language (eg. `CallType` corresponds to `$Call`).
For the rest go to 4.
4. Return `BadEval` exception.

The following constructors will fail step 3:
* `RestType (Object.Rest.ReactConfigMerge, _)`
* `Bind`
* `SpreadType`

```
type Q = $NonMaybeType<?number>;
     ^
```
Here we'll get: `type Q = number"`.

But for:
```
type QP<X> = $NonMaybeType<X>;
     ^
```
We'll get `type QP<X> = $NonMaybeType<X>`.

We cache the results of `EvalT` computation (similarly to tvar caching) for two reasons:
1. [speed] Avoid re-running callbacks to Flow_js
2. [correctness] It turns out that calls to `Flow_js.eval_type_destructor` are not idempotent:
in particular a case that throws exception in the first place will return Empty in subsequent
calls. (A relevant example has been added.)

Reviewed By: samwgoldman

Differential Revision: D8552446

fbshipit-source-id: 35616785493f7a86a89361e77fee8475cd5204be