Disallow `a && b` when `a` is a `number`
Summary:
`0` and `NaN` are both falsy values, so if `a` is `0` (resp. `NaN`), then `a && b` is also `0` (resp. `NaN`).

In many cases, this is not the desired behavior, and JS devs should be using `if`, ternary `?:`, `??`, etc. rather than the `&&` hack. Therefore, let's disallow `number` as the LHS of `&&`.

Reviewed By: gabelevi

Differential Revision: D8592320

fbshipit-source-id: fc4aa132f3df93074ec8158afbbd3fe62b2766ce