Use AST differ for rename
Summary:
This integrates the AST differ into global rename, so that we can give out minimal diffs rather than reprinting the entire file.

Due to an incomplete implementation, the AST differ currently cannot recurse all the way into expressions and therefore ends up creating diffs which are still more granular than they should be here. However, it will be straightforward to add cases to the differ which will improve this.

Reviewed By: panagosg7

Differential Revision: D8706393

fbshipit-source-id: 76a8ecd6dfa5c98c068c5b08dd13ce79e7507c08