Don't write to heaps from rechecker.ml
Summary:
I noticed that `rechecker.ml` was unnecessarily writing to the parsing heaps when it was processing file watcher updates. In order to tell whether Flow should restart after a `package.json` or libdef changes, it parses those files into the heaps and then reads the heaps.

I'm trying to crack down on random code writing to the heaps, so I've replaced this code with code that still reads the files but doesn't mess with the heaps.

Reviewed By: nmote

Differential Revision: D8577075

fbshipit-source-id: 9e4310393d55261e82d329d2cdc840f2d1a83bb3