Diff at the expression level
Summary:
This is fairly straightforward, except that we have to wrap expression in
parentheses to avoid potentially changing the meaning of the program. There is a
test in the next diff which illustrates this danger. There are also many more
details in `js_layout_generator.ml` for anyone curious. Follow the `ctxt`
argument to `Js_layout_generator.expression`.

At some point it would be nice to be able to pass an appropriate context to
`Js_layout_generator.expression`.

Reviewed By: panagosg7

Differential Revision: D8791263

fbshipit-source-id: 4a34ae52c813a55710a856d7252c8420ba8292be