[PR] Create an explicit type for Contexts
Summary:
Sometimes you want to interact with a React Context in an abstract way, such as an implementation for a higher order component that spreads a context consumer into a component.  The way that `createContext` is currently typed makes this a little more difficult than it probably needs to be, so I created a type called `React$Context<T>` and exported the type from the react package as well which seems to be the standard approach for types like this.
Pull Request resolved: https://github.com/facebook/flow/pull/6584

Reviewed By: samwgoldman

Differential Revision: D8804511

Pulled By: gabelevi

fbshipit-source-id: 474ec98f163c8a432be4411c07ad5d49451de6b3