Fix multi-hop bug with object literal properties
Summary:
In certain cases, we can start with multiple def locs. For example, if the query
location is on an object literal property, we include that property's location
as a def loc (see D8154484).

When finding related defs, we use the first element of the list of def locs to
grab elements from the union-find data structure. However, if there are multiple
elements we need to union them first in order for this to work as desired.

Reviewed By: samwgoldman

Differential Revision: D8810259

fbshipit-source-id: 9f24c37638f05d1b9d7084a6a8ffad8013303c71