Fix server crash on persistent-command unhandled exception
Summary:
Should have been using `try%lwt` rather than `try`...

If one of the handlers in handle_persistent_unsafe threw an exception after yielding, then we weren't catching it before this diff, so the server was crashing.

I've observed this myself in the wild - P59807590. I've also spoken with users who I think observed it too (but it's hard to tell).

Reviewed By: gabelevi

Differential Revision: D8812277

fbshipit-source-id: 0a75261908916da53b776dc43a740d6c7cf6eaea