Profile shared memory heap on recheck
Summary:
We currently profile the shared heap after init. If you run `flow check --pretty --profile tests/tuples`, you'll get a `"memory"` section in your JSON that looks like

```
  "memory":{
    "init_done.heap.size":2194432,
    "init_done.hash_table.used_slots":72,
    "init_done.hash_table.slots":524288,
    "init_done.hash_table.nonempty_slots":72,
    "init_done.dep_table.used_slots":0,
    "init_done.dep_table.slots":131072,
    "init_done.dep_table.nonempty_slots":0
  }
```

This is used for Windtunnel profiling (it's important not to break Windtunnel!) and sent to Scuba. But we don't have any logging for rechecks.

This diff makes a few changes (probably could split into multiple diffs but ¯\_(ツ)_/¯)

* `Profiling_js` now can track the starting value, the "high water value" (the maximum value observed), and the ending value.
* `Profiling_js` now can print a pretty little table for the memory stats
* Rechecks now run an aggressive shared memory garbage collect immediately after finishing the recheck (previously they are run later, in `server.ml`'s main loop, after a command is handled)

Reviewed By: samwgoldman

Differential Revision: D8861066

fbshipit-source-id: 321c7bab8f4c6268c58e793f71d72caefbce261d