[Easy] Add regex flags to ./tool
Summary:
Adds --codeFilter and --messageFilter flags, which take a regular expression. This regular expression
is then used to decide whether or not to include errors in the summary. If one is specified to codeFilter, it
looks at the error's context. If one is specified to messageFilter, it looks at the message itself.

Reviewed By: pakoito

Differential Revision: D8915972

fbshipit-source-id: 5dd0c525a440091534a2de454950d94e9cab229c