Refactor lint suppression logic
Summary:
I squashed a few small changes into this diff, which should be 100% semantics
preserving.

The biggest change is that I clarified how the used/unused suppressions logic
works. Before this change, the error suppressions record had two separate
fields for suppressions and unused suppressions. When we added a suppression,
it was added to both maps.

Error collation folds over the errors, accumulating the error suppression
record. Here the fields are treated differently: the suppressions are constant,
but we remove from the "unused" suppressions. At the end of the fold, whatever
is left in in the unused field becomes a warning.

The insight here is that storing two maps is unnecessary. Before collation
the suppressions map is identical to the unused map. This diff takes the
initial suppressions and passes it as a constant argument to the collation
fold, while simultaneously accumualting the unused suppressions.

Reviewed By: gabelevi

Differential Revision: D8882920

fbshipit-source-id: 2418a6c78778b05f460e535d76de2e83e74e0343