Don't mark suppressions for never-enabled lint errors as used
Summary:
Before this diff, we would consider an error suppression to be used if it
precedes a lint error, even if that lint error was never enabled.

For example, given the following file:

```
// flow
declare var x: ?string;
// $FlowFixMe
if (x) {}
```

We would not create an unused suppression warning for the suppression comment,
even when the `sketchy-null-string` lint rule was never enabled.

After this diff, the suppression comment is correctly considered unused.

Reviewed By: gabelevi

Differential Revision: D8887105

fbshipit-source-id: b8580ecd21e75328cdca89c2137b88cb692a1dfd