Transaction
Summary:
This is a very generic `Transaction` module. It runs a function and after the function finishes it commits the transaction. If the function throws, it rolls back the transaction.

I imagine each transaction having a bunch of mutators. Some will mutate heaps (like `ASTHeap`) and some will mutate global refs (like the `Module_js.all_providers` hashtable). If for some reason an exception is thrown during a recheck or command, this will roll the state back to how it was before.

The one downside is we can't throw away oldified data until after the whole recheck (previously we would do it immediately after we finished a step). I'm not yet sure how this will affect memory usage.

Reviewed By: samwgoldman, pakoito

Differential Revision: D8629775

fbshipit-source-id: f58c691246a1c4f05114cd57dd0f2f785a2c8949