Refactor to avoid rechecking dependency cycle
Summary:
This cycle was a pain to break :(

I am working on making force-recheck be able to cancel a
recheck. However, this leads to a cycle which looks like

`Server -> ServerMonitorListener -> CommandHandler -> Rechecker -> ServerMonitorListener`

This refactor addresses it in a few ways:

The cancelable recheck code is moved to `Rechecker`. Now `Rechecker` has
two main APIs for performing rechecks

* `recheck_single` - will perform a single successful recheck if there's
  anything to recheck.
* `recheck_loop` - will keep performing rechecks while there are things
  to recheck.

`CommandHandler` is moved to its own directory, along with its helper
functions

`ServerMonitorListener` is moved to its own directory and split into two
libraries. `ServerMonitorListener` is the loop that listens to the
monitor. `ServerMonitorListenerState` is the state for that loop, which
can be queried by other modules.

Splitting this diff is rather difficult, I'm afraid. Sorry!

Reviewed By: pakoito

Differential Revision: D8789880

fbshipit-source-id: 932a4195caf9a5450d01e700e7c2ba49f7cf85d1