Handle force-recheck immediately
Summary:
I want `force-recheck` to be able to cancel an on-going recheck. That means we can't wait for the recheck to finish before processing the command. We need to handle it immediately as it appears.

The main downside to handling a command immediately is that `env` isn't available.

One of the slightly tricky parts to this is `flow force-recheck --profile`, which waits until the recheck has finished and then sends back the profiling information. Now that we perform the recheck asynchronously, we need to pass through a callback so that we can respond to the client when the recheck finishes.

Reviewed By: pakoito

Differential Revision: D8789884

fbshipit-source-id: 92b79dcc37e33108ee41de5e07c4b8bae51c7441