Make $Flow$DebugSleep cancelable
Summary:
`$Flow$DebugSleep` is a useful little type. It can cause the typechecker to sleep for N seconds, which is a great way to simulate a long recheck. I want to use this to test cancelable rechecks.

Unfortunately, workers only realize that they need to stop when they access shared memory. This is normally not a problem, but here we're just sleeping for N seconds. So to make this useful for testing, I'm changing this to repeated sleep for 1 second and then check if it should die.

Reviewed By: samwgoldman

Differential Revision: D8789877

fbshipit-source-id: 42a811c8de229ef86567adaf906363ba052729cd